fix: remove old datetime range picker,use new one instead
- some style and function fixed
