+++ src/Masa.Stack.Components/Shared/IntegrationComponents/DateTime/SDateTimeRangePicker.razor.css
@@ -1,13 +1,24 @@
-﻿﻿.datetimerangepicker {
-     min-width: 350px;
-     height: 40px;
- }
-
-.datetimerangepicker .regular--text {
-    white-space: nowrap;
+﻿.date-time-range-picker__content ::deep .m-date-picker-table {
+    height: unset;
+    padding: 0;
 }
 
-.datetimerangepicker .m-divider__wrapper {
-    margin: 8px 4px;
-    display: block !important;
+    .date-time-range-picker__content ::deep .m-date-picker-table tr {
+        height: unset;
+    }
+
+.date-time-range-picker__content ::deep .m-card__actions {
+    padding-bottom: 0;
+}
+
+.datetimetext {
+    width: calc(50% - 20px);
+    padding: 0;
+    font-family: 'Roboto';
+    line-height: 21px;
+    color: #485585;
+    font-size: 14px;
+    display: inline-flex;
+    justify-content: center;
+    text-align: center;
 }
